Migrated RuleRepo to Model
